Adding Refs For Missing Evaluation Files
A few files regarding the Evaluation side of things uploaded in May 2015
were missing to be referenced in the AcceptFramework.csproj file and
therefore not loaded as part of the project when open by Visual Studio.
Note: three other not necessary files were also removed in the Accept
API project.